+++ ext/makiri/rust/src/glue/xml_node/mod.rs
@@ -66,6 +66,8 @@ fn init_read() {
         .expect("#document");
     m.define_method("parent", method!(read::parent, 0))
         .expect("#parent");
+    m.define_method("<=>", method!(read::spaceship, 1))
+        .expect("#<=>");
     for name in ["next", "next_sibling"] {
         m.define_method(name, method!(read::next, 0))
             .expect("#next");